Rename TaskCollection to Collection
