+++ src/Collection/Collection.php
@@ -1,5 +1,5 @@
 <?php
-namespace Robo\TaskCollection;
+namespace Robo\Collection;
 
 use Robo\Result;
 use Robo\Contract\TaskInterface;
@@ -19,7 +19,7 @@ use Robo\Contract\CompletionInterface;
  *
  * ``` php
  * <?php
- * $collection = $this->taskCollection();
+ * $collection = $this->collection();
  * $this->taskFileSystemStack()
  *      ->mkdir('logs')
  *      ->touch('logs/.gitignore')
