+++ src/Task/File/TmpFile.php
@@ -2,9 +2,9 @@
 
 namespace Robo\Task\File;
 
-use Robo\TaskCollection\Collection;
+use Robo\Collection\Collection;
 use Robo\Contract\TransientInterface;
-use Robo\TaskCollection\Transient;
+use Robo\Collection\Transient;
 
 /**
  * Create a temporary directory that is automatically cleaned up
