+++ tests/_helpers/_generated/CliGuyActions.php
@@ -1,4 +1,4 @@
-<?php  //[STAMP] 2022e442cd1c1023775634257a1e60e5
+<?php  //[STAMP] 706991b57b11a703db2451535e37c074
 namespace _generated;
 
 // This class was automatically generated by build task
@@ -988,9 +988,9 @@ trait CliGuyActions
      *
      * @param $dirs
      * @return CleanDir
-     * @see \Codeception\Module\CliHelper::taskCollection()
+     * @see \Codeception\Module\CliHelper::collection()
      */
-    public function taskCollection() {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('taskCollection', func_get_args()));
+    public function collection() {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('collection', func_get_args()));
     }
 }
